+++ packages/analyzer_buffer/lib/analyzer_buffer.dart
@@ -1,2 +1,2 @@
-export 'src/analyzer_buffer.dart' hide parseCode;
+export 'src/analyzer_buffer.dart';
 export 'src/revive.dart';