+++ packages/analyzer_buffer/lib/src/analyzer_buffer.dart
@@ -51,6 +51,345 @@ extension CodeFor2 on DartType {
   }
 }
 
+abstract class _Scope {
+  _Scope({
+    required this.uri,
+    required this.show,
+    required this.hide,
+    required this.exports,
+  });
+
+  final List<String> show;
+  final List<String> hide;
+  final Iterable<_ExportScope> exports;
+  final _NormalizedUri uri;
+
+  bool hasAccessTo(
+    _NormalizedUri? targetUri,
+    String symbol, {
+    Set<_NormalizedUri>? visitedImports,
+  }) {
+    visitedImports ??= {};
+    if (visitedImports.contains(uri)) return false;
+
+    visitedImports.add(uri);
+
+    if (show.isNotEmpty && !show.contains(symbol)) {
+      return false;
+    }
+    if (hide.isNotEmpty && hide.contains(symbol)) {
+      return false;
+    }
+
+    // We assume that the symbol is present if the URI matches
+    if (targetUri == uri) return true;
+    if (exports.any((e) => e.hasAccessTo(targetUri, symbol))) {
+      return true;
+    }
+
+    return false;
+  }
+}
+
+class _ExportScope extends _Scope {
+  _ExportScope({
+    required super.uri,
+    required super.show,
+    required super.hide,
+    required super.exports,
+  });
+
+  @override
+  String toString() {
+    final show = this.show.isNotEmpty ? 'show ${this.show.join(', ')}' : '';
+    final hide = this.hide.isNotEmpty ? 'hide ${this.hide.join(', ')}' : '';
+
+    return "export '$uri'$show$hide;";
+  }
+}
+
+class _ImportScope extends _Scope {
+  _ImportScope({
+    required this.prefix,
+    required super.uri,
+    required super.show,
+    required super.hide,
+    required super.exports,
+  });
+
+  final String? prefix;
+
+  @override
+  String toString() {
+    final prefixStr = prefix != null ? ' as $prefix' : '';
+    final show = this.show.isNotEmpty ? 'show ${this.show.join(', ')}' : '';
+    final hide = this.hide.isNotEmpty ? 'hide ${this.hide.join(', ')}' : '';
+
+    return "import '$uri'$prefixStr$show$hide;";
+  }
+}
+
+class _TargetNamespace {
+  _TargetNamespace(
+    _LibraryAdapter library, {
+    required this.generatedFile,
+  }) : imports = library.importedLibraries;
+
+  _TargetNamespace.empty({
+    required this.generatedFile,
+  }) : imports = [];
+
+  final Iterable<_ImportScope> imports;
+  final List<_ImportScope> syntheticImports = [];
+  final _GeneratedFileLocation generatedFile;
+
+  Iterable<_ImportScope> get allImports => imports.followedBy([
+        // TODO only include if dart:core isn't already imported
+        // dart:core is always imported, so we add it as a synthetic import.
+        _ImportScope(
+          prefix: null,
+          uri: _NormalizedUri._(Uri.parse('dart:core')),
+          show: [],
+          hide: [],
+          exports: [],
+        ),
+
+        // The pack for the current file
+        _ImportScope(
+          prefix: null,
+          uri: _NormalizedUri.fromImportUri(
+            generatedFile: generatedFile,
+            Uri(
+              scheme: 'asset',
+              path: path.joinAll([
+                generatedFile._packageName,
+                path.normalize(generatedFile._path),
+              ]),
+            ),
+          ),
+          show: [],
+          hide: [],
+          exports: [],
+        ),
+      ]).followedBy(syntheticImports);
+
+  _ImportScope? findSymbol(_NormalizedUri uri, String symbol) {
+    return allImports.where((e) => e.hasAccessTo(uri, symbol)).firstOrNull;
+  }
+
+  _ImportScope import(_NormalizedUri uri, String symbol) {
+    final import = findSymbol(uri, symbol);
+    if (import != null) return import;
+
+    final prefix = '_i${imports.length + 1}';
+    final newImport = _ImportScope(
+      uri: uri,
+      prefix: prefix,
+      show: [],
+      hide: [],
+      exports: [],
+    );
+    syntheticImports.add(newImport);
+    return newImport;
+  }
+}
+
+/// Information about the generated file.
+final class _GeneratedFileLocation {
+  /// A manually defined location of the generated file.
+  ///
+  /// [path] should be relative to the package root, such as `lib/generated/file.dart`.
+  /// [packageName] should be the name of the package where the file is located.
+  _GeneratedFileLocation({
+    required String packageName,
+    required String path,
+  })  : _path = path,
+        _packageName = packageName;
+
+  factory _GeneratedFileLocation._relativeTo(Uri uri) {
+    switch (uri) {
+      case Uri(scheme: != 'package' && != 'asset'):
+        throw ArgumentError(
+          'GeneratedFileLocation required a package/asset URI but got: $uri',
+        );
+      case Uri(
+          scheme: 'package',
+          pathSegments: [final packageName, ...final other]
+        ):
+        return _GeneratedFileLocation(
+          packageName: packageName,
+          path: path.joinAll(['lib', ...other]),
+        );
+      case Uri(
+          scheme: 'asset',
+          pathSegments: [final packageName, ...final other]
+        ):
+        return _GeneratedFileLocation(
+          packageName: packageName,
+          path: path.joinAll(other),
+        );
+      case _:
+        throw ArgumentError(
+          'Badly formatted URI: $uri',
+        );
+    }
+  }
+
+  final String _packageName;
+  final String _path;
+
+  late final uri = Uri(
+    scheme: 'asset',
+    path: '$_packageName/${path.normalize(_path)}',
+  );
+}
+
+@immutable
+class _NormalizedUri {
+  const _NormalizedUri._(this.uri);
+
+  factory _NormalizedUri.fromCode(
+    Uri input, {
+    required _GeneratedFileLocation generatedFile,
+  }) {
+    var uri = input;
+    if (!uri.hasScheme) uri = uri.replace(scheme: 'package');
+
+    switch (uri) {
+      case Uri(scheme: 'file' || 'dart'):
+        break;
+      case Uri(scheme: 'asset', pathSegments: [final packageName, ...]):
+        if (packageName != generatedFile._packageName) {
+          throw ArgumentError(
+            'Asset URI $uri does not match the package name ${generatedFile._packageName}',
+          );
+        }
+
+      case Uri(scheme: 'package', pathSegments: [final segment]):
+        uri = uri.replace(pathSegments: [segment, '$segment.dart']);
+      case Uri(scheme: 'package', pathSegments: [...final rest, final last])
+          when !last.endsWith('.dart'):
+        uri = uri.replace(pathSegments: [...rest, '$last.dart']);
+      case Uri(scheme: 'package'):
+        break;
+      case _:
+        throw UnsupportedError('Unsupported URI: $uri');
+    }
+
+    return _NormalizedUri.fromImportUri(
+      uri,
+      generatedFile: generatedFile,
+    );
+  }
+
+  factory _NormalizedUri.fromLibrary(
+    LibraryElement element, {
+    required _GeneratedFileLocation generatedFile,
+  }) {
+    final uri = element.source.uri;
+    if (uri.scheme != 'package' &&
+        uri.scheme != 'asset' &&
+        uri.scheme != 'dart') {
+      throw ArgumentError(
+        'Expected a package/asset/dart URI, but got: $uri',
+      );
+    }
+
+    return _NormalizedUri.fromImportUri(
+      uri,
+      generatedFile: generatedFile,
+    );
+  }
+
+  factory _NormalizedUri.fromLibrary2(
+    LibraryElement2 element, {
+    required _GeneratedFileLocation generatedFile,
+  }) {
+    final uri = element.uri;
+    if (uri.scheme != 'package' &&
+        uri.scheme != 'asset' &&
+        uri.scheme != 'dart') {
+      throw ArgumentError(
+        'Expected a package/asset/dart URI, but got: $uri',
+      );
+    }
+
+    return _NormalizedUri.fromImportUri(
+      uri,
+      generatedFile: generatedFile,
+    );
+  }
+
+  factory _NormalizedUri.fromImportUri(
+    Uri uri, {
+    required _GeneratedFileLocation generatedFile,
+  }) {
+    switch (uri) {
+      // import 'dir/file.dart' -> 'asset:package_name/path_to_generated/../dir/file.dart'
+      case Uri(hasScheme: false):
+        uri = uri.replace(
+          scheme: 'asset',
+          path: path.join(path.dirname(generatedFile._path), uri.path),
+        );
+      // import 'package:package_name/dir/file.dart' -> 'asset:package_name/lib/dir/file.dart'
+      case Uri(
+            scheme: 'package',
+            pathSegments: [final packageName, ...final rest]
+          )
+          when packageName == generatedFile._packageName:
+        uri = uri.replace(
+          scheme: 'asset',
+          path: path.joinAll([packageName, 'lib', ...rest]),
+        );
+    }
+
+    return _NormalizedUri._(uri.replace(path: path.normalize(uri.path)));
+  }
+
+  Uri toImportUri({
+    required _GeneratedFileLocation generatedFile,
+  }) {
+    switch (uri) {
+      // asset:package_name/lib/dir/file.dart -> ./dir/file.dart
+      case Uri(
+          scheme: 'asset',
+          pathSegments: [final packageName, ...final rest]
+        ):
+        if (packageName != generatedFile._packageName) {
+          throw ArgumentError(
+            'Asset URI $uri does not match the package name ${generatedFile._packageName}',
+          );
+        }
+        return Uri.parse(
+          path.normalize(
+            path.relative(
+              path.joinAll(rest),
+              from: path.dirname(generatedFile._path),
+            ),
+          ),
+        );
+      case _:
+        return uri;
+    }
+  }
+
+  final Uri uri;
+
+  @override
+  bool operator ==(Object other) {
+    if (other is! _NormalizedUri) return false;
+    return uri == other.uri;
+  }
+
+  @override
+  int get hashCode => uri.hashCode;
+
+  @override
+  String toString() {
+    return '$_NormalizedUri($uri)';
+  }
+}
+
 /// A String buffer that is used to generate Dart code.
 ///
 /// It helps deal with common problems such as:
@@ -63,14 +402,36 @@ extension CodeFor2 on DartType {
 /// Alternatively, you can write code "as usual" by using the [writeType] method,
 /// combined with passing non-interpolated strings to [write].
 class AnalyzerBuffer {
+  AnalyzerBuffer._({
+    required _GeneratedFileLocation generatedFile,
+    required _TargetNamespace namespace,
+    required this.header,
+    required bool autoImport,
+  })  : _generatedFile = generatedFile,
+        _namespace = namespace,
+        _autoImport = autoImport;
+
   /// Creates a [AnalyzerBuffer] that generates a brand new library.
   ///
   /// When writing types, [AnalyzerBuffer] will automatically add the necessary imports
   /// to the generated code.
+  ///
+  /// [packageName] is the name of the package where the generated file will be located.
+  /// [path] is the path of the generated file, relative to the package root (e.g. `lib/src/file.dart`).
   AnalyzerBuffer.newLibrary({
-    required this.sourcePath,
+    required String packageName,
+    required String path,
     this.header,
-  })  : _libraryAdapter = null,
+  })  : _generatedFile = _GeneratedFileLocation(
+          packageName: packageName,
+          path: path,
+        ),
+        _namespace = _TargetNamespace.empty(
+          generatedFile: _GeneratedFileLocation(
+            packageName: packageName,
+            path: path,
+          ),
+        ),
         _autoImport = true;
 
   /// Creates a [AnalyzerBuffer] that generates code for a specific [library].
@@ -78,27 +439,51 @@ class AnalyzerBuffer {
   /// This will not automatically import missing libraries.
   /// Instead, it will rely on the existing imports to decide which prefix to use
   /// when writing types.
-  AnalyzerBuffer.fromLibrary(
+  factory AnalyzerBuffer.part(
     LibraryElement library, {
-    required this.sourcePath,
-    this.header,
-  })  : _libraryAdapter = _LegacyLibraryAdapter(library),
-        _autoImport = false;
+    String? header,
+  }) {
+    final generatedFiles =
+        _GeneratedFileLocation._relativeTo(library.source.uri);
+
+    final namespace = _TargetNamespace(
+      _LegacyLibraryAdapter(library, generatedFiles),
+      generatedFile: generatedFiles,
+    );
+
+    return AnalyzerBuffer._(
+      generatedFile: generatedFiles,
+      namespace: namespace,
+      header: header,
+      autoImport: false,
+    );
+  }
 
   /// Creates a [AnalyzerBuffer] that generates code for a specific [library].
   ///
   /// This will not automatically import missing libraries.
   /// Instead, it will rely on the existing imports to decide which prefix to use
   /// when writing types.
-  AnalyzerBuffer.fromLibrary2(
+  factory AnalyzerBuffer.part2(
     LibraryElement2 library, {
-    required this.sourcePath,
-    this.header,
-  })  : _libraryAdapter = _LibraryAdapter2(library),
-        _autoImport = false;
+    String? header,
+  }) {
+    final generatedFiles = _GeneratedFileLocation._relativeTo(library.uri);
 
-  /// The associated library, if any.
-  final _LibraryAdapter? _libraryAdapter;
+    final namespace = _TargetNamespace(
+      _LibraryAdapter2(library, generatedFiles),
+      generatedFile: generatedFiles,
+    );
+
+    return AnalyzerBuffer._(
+      generatedFile: generatedFiles,
+      namespace: namespace,
+      header: header,
+      autoImport: false,
+    );
+  }
+
+  final _TargetNamespace _namespace;
   final StringBuffer _buffer = StringBuffer();
 
   /// Whether any write was performed on the buffer.
@@ -110,45 +495,23 @@ class AnalyzerBuffer {
   final bool _autoImport;
 
   /// The path of the file that is being generated.
-  final String sourcePath;
+  final _GeneratedFileLocation _generatedFile;
 
   /// A header that will be added at the top of the generated code.
   final String? header;
 
   void Function()? Function(String name)? _lookupArg;
 
-  Iterable<_LibraryAdapter> get _currentlyImportedLibraries {
-    Iterable<_LibraryAdapter> result = _importedLibraries;
-    if (_libraryAdapter case final library?) {
-      result = result.followedBy(
-        library.importedLibraries,
-      );
-    }
-    return result;
-  }
+  _ImportScope _upsertImport(_NormalizedUri uri, String symbol) {
+    final find = _namespace.findSymbol(uri, symbol);
+    if (find != null) return find;
 
-  final List<_LibraryAdapter> _importedLibraries = [];
-  String? _importLibrary(_LibraryAdapter library) {
-    // final library = adapter.library2;
-    // // dart:core have a null library, so we don't import it.
-    // if (library == null) return null;
+    if (_autoImport) return _namespace.import(uri, symbol);
 
-    if (library.uri case Uri(scheme: 'dart', path: 'core')) return null;
-
-    if (!_importedLibraries.contains(library)) _importedLibraries.add(library);
-
-    return _prefixFor(library);
-  }
-
-  String? _prefixFor(_LibraryAdapter other) {
-    final index = _importedLibraries.indexOf(other);
-    if (index >= 0) return '_i${index + 1}';
-
-    if (_libraryAdapter case final library?) {
-      return library.findPrefixFor(other.uri);
-    }
-
-    return null;
+    throw ArgumentError(
+      'Cannot find import for $symbol in $uri, and could not automatically import it.\n'
+      'Available imports:\n${_namespace.allImports.map((e) => e.uri.toString()).join('\n')}',
+    );
   }
 
   /// Writes the given [type] to the buffer.
@@ -164,14 +527,11 @@ class AnalyzerBuffer {
     bool recursive = true,
   }) {
     type._visit(
+      generatedFile: _generatedFile,
       onType: (library, name, suffix, args) {
         String? prefix;
         if (library != null) {
-          if (_autoImport) {
-            prefix = _importLibrary(library);
-          } else {
-            prefix = _prefixFor(library);
-          }
+          prefix = _upsertImport(library.uri, name).prefix;
         }
 
         if (prefix != null) {
@@ -288,9 +648,9 @@ class AnalyzerBuffer {
         previousIndex = match.end;
 
         final matchedString = match.group(1)!;
-        parseCode(
+        _parseCode(
           matchedString,
-          sourcePath: sourcePath,
+          generatedFile: _generatedFile,
           onArg: (argName) {
             final arg = lookup(argName);
             if (arg == null) {
@@ -298,14 +658,13 @@ class AnalyzerBuffer {
             }
             arg();
           },
-          onUri: (uri, type) {
-            final prefix =
-                _autoImport ? _importLibraryUri(uri) : _prefixForUri(uri);
+          onUri: (uri, symbol) {
+            final prefix = _upsertImport(uri, symbol).prefix;
             if (prefix != null) {
               _buffer.write(prefix);
               _buffer.write('.');
             }
-            _buffer.write(type);
+            _buffer.write(symbol);
           },
         );
       }
@@ -316,36 +675,6 @@ class AnalyzerBuffer {
     }
   }
 
-  final List<Uri> _importedLibraryUris = [];
-  String? _importLibraryUri(Uri uri) {
-    if (uri.isIgnoredFromImports) return null;
-
-    final alreadyImportedLibrary =
-        _currentlyImportedLibraries.where((e) => e.uri == uri).firstOrNull;
-    if (alreadyImportedLibrary != null) {
-      return _prefixFor(alreadyImportedLibrary);
-    }
-
-    final index = _importedLibraryUris.indexOf(uri);
-    if (index >= 0) return _prefixForUri(uri);
-
-    _importedLibraryUris.add(uri);
-    return _prefixForUri(uri);
-  }
-
-  String? _prefixForUri(Uri uri) {
-    if (uri.isIgnoredFromImports) return null;
-
-    final index = _importedLibraryUris.indexOf(uri);
-    if (index >= 0) return '_u${index + 1}';
-
-    if (_libraryAdapter case final library?) {
-      return library.findPrefixFor(uri);
-    }
-
-    return null;
-  }
-
   @override
   String toString() {
     if (isEmpty) return '';
@@ -353,56 +682,31 @@ class AnalyzerBuffer {
     return [
       '// GENERATED CODE - DO NOT MODIFY BY HAND',
       if (header != null) header,
-      ..._importedLibraries.mapIndexed(
-        (index, e) {
-          final prefix = _prefixFor(e);
-          if (prefix == null) return "import '${e.uri}';";
-          return "import '${e.uri}' as $prefix;";
-        },
-      ),
-      ..._importedLibraryUris.mapIndexed(
-        (index, e) {
-          final prefix = _prefixForUri(e);
-          if (prefix == null) return "import '$e';";
-          return "import '$e' as $prefix;";
-        },
-      ),
+      ..._namespace.syntheticImports.map((e) {
+        final prefix = e.prefix;
+        final targetUri = e.uri.toImportUri(generatedFile: _generatedFile);
+
+        if (prefix == null) return "import '$targetUri';";
+        return "import '$targetUri' as $prefix;";
+      }),
       _buffer,
     ].join('\n');
   }
 }
 
-extension on Uri {
-  bool get isIgnoredFromImports {
-    switch (this) {
-      // Don't import dart:core, as it is the default library.
-      case Uri(scheme: 'dart', path: 'core'):
-      // Don't import self.
-      case Uri(scheme: '', pathSegments: []):
-      case Uri(scheme: '', pathSegments: ['']):
-      case Uri(scheme: '', pathSegments: ['.']):
-      case Uri(scheme: '', pathSegments: ['./']):
-        return true;
-    }
-
-    return false;
-  }
-}
-
-@internal
-void parseCode(
+void _parseCode(
   String code, {
   required void Function(String arg) onArg,
-  required void Function(Uri uri, String type) onUri,
-  required String sourcePath,
+  required void Function(_NormalizedUri uri, String type) onUri,
+  required _GeneratedFileLocation generatedFile,
 }) {
   switch (code.split('|')) {
     case [final argName]:
       onArg(argName);
     case [final uriStr, final type]:
-      final uri = _normalizeCodeUri(
+      final uri = _NormalizedUri.fromCode(
         Uri.parse(uriStr),
-        sourcePath: sourcePath,
+        generatedFile: generatedFile,
       );
 
       onUri(uri, type);
@@ -411,45 +715,13 @@ void parseCode(
   }
 }
 
-Uri _normalizeCodeUri(
-  Uri input, {
-  required String sourcePath,
-}) {
-  var uri = input;
-  if (!uri.hasScheme) uri = uri.replace(scheme: 'package');
-
-  switch (uri) {
-    case Uri(scheme: 'file' || 'dart'):
-      break;
-    case Uri(
-        scheme: 'asset',
-        pathSegments: [_ /* package name */, ...final pathSegments]
-      ):
-      final targetPath = path.joinAll(pathSegments);
-      uri = uri.replace(
-        scheme: '',
-        path: path.normalize(
-          path.relative(targetPath, from: path.dirname(sourcePath)),
-        ),
-      );
-    case Uri(scheme: 'package', pathSegments: [final segment]):
-      uri = uri.replace(pathSegments: [segment, '$segment.dart']);
-    case Uri(scheme: 'package', pathSegments: [...final rest, final last])
-        when !last.endsWith('.dart'):
-      uri = uri.replace(pathSegments: [...rest, '$last.dart']);
-    case Uri(scheme: 'package'):
-      break;
-    case _:
-      throw UnsupportedError('Unsupported URI: $uri');
-  }
-
-  return uri;
-}
-
 extension on Element2 {
-  _LibraryAdapter2? get libraryAdapter => switch (library2) {
+  _LibraryAdapter2? libraryAdapter({
+    required _GeneratedFileLocation generatedFile,
+  }) =>
+      switch (library2) {
         null => null,
-        final LibraryElement2 e => _LibraryAdapter2(e),
+        final LibraryElement2 e => _LibraryAdapter2(e, generatedFile),
       };
 }
 
@@ -462,11 +734,14 @@ extension on DartType {
       List<DartType> args,
     ) onType,
     required void Function(RecordType type) onRecord,
+    required _GeneratedFileLocation generatedFile,
   }) {
     final alias = this.alias;
     if (alias != null) {
       onType(
-        alias.element2.libraryAdapter,
+        alias.element2.libraryAdapter(
+          generatedFile: generatedFile,
+        ),
         alias.element2.name3!,
         nullabilitySuffix,
         alias.typeArguments,
@@ -489,7 +764,9 @@ extension on DartType {
 
     if (that is ParameterizedType) {
       onType(
-        that.element3?.libraryAdapter,
+        that.element3?.libraryAdapter(
+          generatedFile: generatedFile,
+        ),
         name,
         nullabilitySuffix,
         that.typeArguments,
@@ -498,7 +775,9 @@ extension on DartType {
     }
 
     onType(
-      that.element3?.libraryAdapter,
+      that.element3?.libraryAdapter(
+        generatedFile: generatedFile,
+      ),
       name,
       nullabilitySuffix,
       [],
@@ -508,11 +787,9 @@ extension on DartType {
 
 @immutable
 abstract base class _LibraryAdapter<T extends _LibraryAdapter<T>> {
-  Uri get uri;
-
-  String? findPrefixFor(Uri uri);
-
-  Iterable<_LibraryAdapter<T>> get importedLibraries;
+  _NormalizedUri get uri;
+  Iterable<_ImportScope> get importedLibraries;
+  Iterable<_ExportScope> get exportedLibraries;
 
   @override
   bool operator ==(Object other) =>
@@ -523,76 +800,142 @@ abstract base class _LibraryAdapter<T extends _LibraryAdapter<T>> {
 
   @override
   String toString() {
-    return 'LibraryAdapter(uri: $uri, importedLibraries: ${importedLibraries.toList()})';
+    return '$runtimeType($uri})';
   }
 }
 
 final class _LegacyLibraryAdapter
     extends _LibraryAdapter<_LegacyLibraryAdapter> {
-  _LegacyLibraryAdapter(this.library);
+  _LegacyLibraryAdapter(this.library, this.generatedFile);
+
   final LibraryElement library;
+  final _GeneratedFileLocation generatedFile;
 
   @override
-  Uri get uri => library.source.uri;
+  _NormalizedUri get uri => _NormalizedUri.fromLibrary(
+        library,
+        generatedFile: generatedFile,
+      );
 
   @override
-  Iterable<_LegacyLibraryAdapter> get importedLibraries {
-    return library.importedLibraries.map(_LegacyLibraryAdapter.new);
+  Iterable<_ImportScope> get importedLibraries sync* {
+    for (final importedLibrary in library.importedLibraries) {
+      final import = library.definingCompilationUnit.libraryImportPrefixes
+          .expand((e) => e.imports)
+          .where(
+            (e) => e.importedLibrary?.source.uri == importedLibrary.source.uri,
+          )
+          .firstOrNull;
+
+      yield _ImportScope(
+        uri: _NormalizedUri.fromLibrary(
+          importedLibrary,
+          generatedFile: generatedFile,
+        ),
+        prefix: import?.prefix?.element.name,
+        show: import?.combinators
+                .whereType<ShowElementCombinator>()
+                .expand((e) => e.shownNames)
+                .toList() ??
+            [],
+        hide: import?.combinators
+                .whereType<HideElementCombinator>()
+                .expand((e) => e.hiddenNames)
+                .toList() ??
+            [],
+        exports: _LegacyLibraryAdapter(importedLibrary, generatedFile)
+            .exportedLibraries,
+      );
+    }
   }
 
   @override
-  String? findPrefixFor(Uri uri) {
-    final prefix = library.definingCompilationUnit.libraryImportPrefixes
-        .expand((e) => e.imports)
-        .where((e) {
-          final importedLibrary = e.importedLibrary;
-          if (importedLibrary == null) return false;
-          if (importedLibrary.source.uri == uri) return true;
-          for (final lib in importedLibrary.exportedLibraries) {
-            if (lib.source.uri == uri) return true;
-          }
+  Iterable<_ExportScope> get exportedLibraries sync* {
+    for (final export in library.definingCompilationUnit.libraryExports) {
+      final exportedLibrary = export.exportedLibrary;
+      if (exportedLibrary == null) continue;
 
-          return false;
-        })
-        .firstOrNull
-        ?.prefix;
-
-    return prefix?.element.name;
+      yield _ExportScope(
+        uri: _NormalizedUri.fromLibrary(
+          exportedLibrary,
+          generatedFile: generatedFile,
+        ),
+        show: export.combinators
+            .whereType<ShowElementCombinator>()
+            .expand((e) => e.shownNames)
+            .toList(),
+        hide: export.combinators
+            .whereType<HideElementCombinator>()
+            .expand((e) => e.hiddenNames)
+            .toList(),
+        exports: _LegacyLibraryAdapter(exportedLibrary, generatedFile)
+            .exportedLibraries,
+      );
+    }
   }
 }
 
 final class _LibraryAdapter2 extends _LibraryAdapter<_LibraryAdapter2> {
-  _LibraryAdapter2(this.library);
+  _LibraryAdapter2(this.library, this.generatedFile);
   final LibraryElement2 library;
+  final _GeneratedFileLocation generatedFile;
 
   @override
-  Uri get uri => library.uri;
+  _NormalizedUri get uri =>
+      _NormalizedUri.fromLibrary2(library, generatedFile: generatedFile);
 
   @override
-  Iterable<_LibraryAdapter2> get importedLibraries {
-    return library.fragments
-        .expand((e) => e.importedLibraries2)
-        .map(_LibraryAdapter2.new);
+  Iterable<_ImportScope> get importedLibraries sync* {
+    for (final importPrefix in library.fragments) {
+      for (final import in importPrefix.libraryImports2) {
+        final importedLibrary2 = import.importedLibrary2;
+        if (importedLibrary2 == null) continue;
+
+        yield _ImportScope(
+          uri: _NormalizedUri.fromLibrary2(
+            importedLibrary2,
+            generatedFile: generatedFile,
+          ),
+          prefix: import.prefix2?.element.name3,
+          show: import.combinators
+              .whereType<ShowElementCombinator>()
+              .expand((e) => e.shownNames)
+              .toList(),
+          hide: import.combinators
+              .whereType<HideElementCombinator>()
+              .expand((e) => e.hiddenNames)
+              .toList(),
+          exports: _LibraryAdapter2(importedLibrary2, generatedFile)
+              .exportedLibraries,
+        );
+      }
+    }
   }
 
   @override
-  String? findPrefixFor(Uri uri) {
-    final prefix = library.fragments
-        .expand((e) => e.prefixes)
-        .expand((e) => e.imports)
-        .where((e) {
-          final importedLibrary2 = e.importedLibrary2;
-          if (importedLibrary2 == null) return false;
-          if (importedLibrary2.uri == uri) return true;
-          for (final lib in importedLibrary2.exportedLibraries2) {
-            if (lib.uri == uri) return true;
-          }
+  Iterable<_ExportScope> get exportedLibraries sync* {
+    for (final export in library.fragments) {
+      for (final exportedLibrary in export.libraryExports2) {
+        final exportedLibrary2 = exportedLibrary.exportedLibrary2;
+        if (exportedLibrary2 == null) continue;
 
-          return false;
-        })
-        .firstOrNull
-        ?.prefix2;
-
-    return prefix?.element.name3;
+        yield _ExportScope(
+          uri: _NormalizedUri.fromLibrary2(
+            exportedLibrary2,
+            generatedFile: generatedFile,
+          ),
+          show: exportedLibrary.combinators
+              .whereType<ShowElementCombinator>()
+              .expand((e) => e.shownNames)
+              .toList(),
+          hide: exportedLibrary.combinators
+              .whereType<HideElementCombinator>()
+              .expand((e) => e.hiddenNames)
+              .toList(),
+          exports: _LibraryAdapter2(exportedLibrary2, generatedFile)
+              .exportedLibraries,
+        );
+      }
+    }
   }
 }