+++ packages/e2e/lib/builder.dart
@@ -5,18 +5,38 @@ import 'dart:async';
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer_buffer/analyzer_buffer.dart';
 import 'package:build/build.dart';
+import 'package:path/path.dart' as path;
 import 'package:source_gen/source_gen.dart';
 import './annotation.dart';
 
-/// Builds generators for `build_runner` to run
-Builder e2e(BuilderOptions options) {
+Builder partBuilder(BuilderOptions options) {
   return SharedPartBuilder(
     [const E2EPart()],
     'e2e',
   );
 }
 
-Builder e2e2(BuilderOptions options) {
+class E2EPart extends GeneratorForAnnotation<E2E> {
+  const E2EPart();
+
+  @override
+  Future<String> generateForAnnotatedElement(
+    Element element,
+    ConstantReader annotation,
+    BuildStep buildStep,
+  ) async {
+    final buffer = AnalyzerBuffer.part(
+      element.library!,
+      header: '// ignore_for_file: type=lint, type=warning',
+    );
+
+    _handle(element, buffer);
+
+    return buffer.toString();
+  }
+}
+
+Builder libBuilder(BuilderOptions options) {
   return LibraryBuilder(
     E2ELibrary(),
     generatedExtension: '.e2e.dart',
@@ -29,7 +49,11 @@ class E2ELibrary extends Generator {
   @override
   FutureOr<String?> generate(LibraryReader library, BuildStep buildStep) {
     final buffer = AnalyzerBuffer.newLibrary(
-      sourcePath: buildStep.inputId.path,
+      packageName: buildStep.inputId.package,
+      path: path.join(
+        path.dirname(buildStep.inputId.path),
+        '${path.basenameWithoutExtension(buildStep.inputId.path)}.e2e.dart',
+      ),
       header: '// ignore_for_file: type=lint, type=warning',
     );
 
@@ -46,27 +70,6 @@ class E2ELibrary extends Generator {
   }
 }
 
-class E2EPart extends GeneratorForAnnotation<E2E> {
-  const E2EPart();
-
-  @override
-  Future<String> generateForAnnotatedElement(
-    Element element,
-    ConstantReader annotation,
-    BuildStep buildStep,
-  ) async {
-    final buffer = AnalyzerBuffer.fromLibrary(
-      element.library!,
-      sourcePath: buildStep.inputId.path,
-      header: '// ignore_for_file: type=lint, type=warning',
-    );
-
-    _handle(element, buffer);
-
-    return buffer.toString();
-  }
-}
-
 void _handle(Element element, AnalyzerBuffer buffer) {
   switch (element) {
     case FunctionElement():
@@ -74,7 +77,7 @@ void _handle(Element element, AnalyzerBuffer buffer) {
       buffer.write(' ${element.name}E2e([');
       for (final parameter in element.parameters) {
         buffer.write(
-          '${parameter.type.toCode()} ${element.name}',
+          '${parameter.type.toCode()} ${parameter.name}',
         );
         if (parameter.hasDefaultValue) {
           buffer.write(' = ${parameter.computeConstantValue()!.toCode()}');