+++ packages/e2e/test/integration/first.g.dart
@@ -8,7 +8,17 @@ part of 'first.dart';
 
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint, type=warning
-void fnE2e([
-  Object fn = const Foo(),
+void simpleFromTestE2e([
+  Object value = const Foo(),
+]) =>
+    throw UnimplementedError();
+
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint, type=warning
+void prefixToRelativePathE2e([
+  Object value = const prefix.A(),
+  Object value2 = const not_normalized_path_prefix.B(),
+  Object value3 = const direct_path_prefix.C(),
+  Object value6 = const relative_reexport.F(),
 ]) =>
     throw UnimplementedError();